Make sure keys exist before trying to use them

+++ scripts/php7dev.rb
@@ -36,18 +36,21 @@ class Php7dev
     
     if !Vagrant::Util::Platform.windows?
       # Configure The Public Key For SSH Access
-      config.vm.provision "shell" do |s|
-        s.inline = "echo $1 | grep -xq \"$1\" /home/vagrant/.ssh/authorized_keys || echo $1 | tee -a /home/vagrant/.ssh/authorized_keys"
-        s.args = [File.read(File.expand_path(settings["authorize"]))]
+      if File.exists? File.expand_path(settings["authorize"]) then
+        config.vm.provision "shell" do |s|
+          s.inline = "echo $1 | grep -xq \"$1\" /home/vagrant/.ssh/authorized_keys || echo $1 | tee -a /home/vagrant/.ssh/authorized_keys"
+          s.args = [File.read(File.expand_path(settings["authorize"]))]
+        end
       end
-
       # Copy The SSH Private Keys To The Box
       settings["keys"].each do |key|
-        config.vm.provision "shell" do |s|
-          s.privileged = false
-          s.inline = "echo \"$1\" > /home/vagrant/.ssh/$2 && chmod 600 /home/vagrant/.ssh/$2"
-          s.args = [File.read(File.expand_path(key)), key.split('/').last]
-       end
+        if File.exists? File.expand_path(key) then
+          config.vm.provision "shell" do |s|
+            s.privileged = false
+            s.inline = "echo \"$1\" > /home/vagrant/.ssh/$2 && chmod 600 /home/vagrant/.ssh/$2"
+            s.args = [File.read(File.expand_path(key)), key.split('/').last]
+         end
+        end
       end
     end
